requirements.txt
future
six
smbus2
RPi.GPIO
